requirements_2.txt
pip==9.0.1
setuptools==36.0.1
alabaster==0.7.10
alembic==0.9.2
amqp==2.1.4
anyjson==0.3.3
anykeystore==0.2
argh==0.26.2
# Do not specify argparse, jenkins chokes on system version.
# argparse==1.2.2
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
-e git+https://github.com/conversence/Beaker.git@assembl#egg=Beaker
beautifulsoup4==4.6.0
billiard==3.5.0.2
boto==2.47.0
bz2file==0.98
celery==4.0.2
cffi==1.10.0
Chameleon==3.1
colander==1.3.3
colors.py==0.2.2
contextlib2==0.5.5
cornice==2.4.0
-e git+https://github.com/maparent/cornice.ext.sphinx.git@cornice_sphinx_services#egg=cornice_sphinx
cov-core==1.15.0
coverage==4.4.1
cryptography==1.9
cssselect==1.0.1
cssutils==1.0.2
DateTime==4.2
defusedxml==0.4.1  # rq.filter: <0.5
# distribute==0.7.3 # This fails with latest pip. Installed properly as a dependency.
# constrained by saml
dm.xmlsec.binding==1.3.3
docutils==0.13.1
dogpile.cache==0.6.4
dogpile.core==0.4.1
ecdsa==0.13
enum34==1.1.6
et-xmlfile==1.0.1
Fabric==1.13.2
facebook-sdk==1.0.0  # rq.filter: <2
feedparser==5.2.1
flaky==3.4.0
#Because mock 2.0.0 requires it, but it isn't always detected properly
funcsigs==1.0.2
future==0.16.0
-e git+https://github.com/maparent/fuxi.git#egg=fuxi
gensim==2.2.0
google-api-python-client==1.6.2
greenlet==0.4.12
hiredis==0.2.0
html5lib==0.999999999
httplib2==0.10.3
hupper==1.0
idna==2.5
imagesize==0.7.1
#IMAPClient==1.0.1
imaplib2==2.45.0
ipaddress==1.0.18
ipykernel==4.6.1
ipython-genutils==0.2.0
ipython==5.4.1 # rq.filter: <6
iso639==0.1.4  # failure with 0.1.3
isodate==0.5.4
-e git+https://github.com/assembl/jasmine-splinter-runner.git#egg=jasmine-splinter-runner
jdcal==1.3
Jinja2==2.9.6
jupyter-client==5.1.0
jupyter-core==4.3.0
-e git+https://github.com/benoitg/jwzthreading.git@master#egg=jwzthreading
kombu==4.0.2
langdetect==1.0.7
lingua==4.13
lxml==3.8.0
Mako==1.0.6
MarkupSafe==1.0
meld3==1.0.2
mock==2.0.0
ndg_httpsclient==0.4.2
nodeenv==1.1.4
numpy==1.13.0
#oauthlib==2.0.2
oauth2client==4.1.2
openpyxl==2.4.8
paramiko==2.2.1
PasteDeploy==1.5.2
pathlib2==2.3.0
pathtools==0.1.2
PdbSublimeTextSupport==0.2
pexpect==4.2.1
pickleshare==0.7.4
plaster-pastedeploy==0.3.1
plaster==0.5
polib==1.0.8
premailer==3.0.1
psycopg2==2.7.1
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.2.3
pyasn1_modules==0.0.9
pycparser==2.17
pycryptopp==0.7.1.869544967005693312591928092448767568728501330214
pydot==1.2.3
Pygments==2.2.0
pygraphviz==1.3.1
pyIsEmail==v1.3.1
dnspython==1.15.0   # must come after pyisemail?
PyJWT==1.5.2
pylibmc==1.5.2
pyOpenSSL==17.0.0
pyparsing==2.2.0
# Until https://github.com/SurveyMonkey/pyramid_autodoc/pull/12 is merged
-e git+https://github.com/matthewnorman/pyramid_autodoc@sphinx_161_no_domains_v2#egg=pyramid_autodoc
-e git+https://github.com/assembl/pyramid_beaker.git@assembl#egg=pyramid_beaker
pyramid-debugtoolbar-ajax==0.1.0
pyramid-debugtoolbar==4.2
pyramid-dogpile-cache==0.0.4
pyramid-ipython==0.2
pyramid-jinja2==2.7
pyramid-mailer==0.15.1
pyramid-mako==1.0.2
pyramid-tm==2.1
pyramid==1.9
PyStemmer==1.3.0
pytest-cov==2.5.1
pytest-localserver==0.3.7
pytest-mock==1.6.0
pytest-splinter==1.8.5
pytest==3.1.2
python-dateutil==2.6.0
python-openid==2.2.5
python-saml==2.2.3
python-snappy==0.5.1
-e git+https://github.com/assembl/python-social-auth.git@assembl#egg=python_social_auth
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
raven==6.1.0
rdflib-jsonld==0.4.0
rdflib==4.2.2
reg==0.11
redis==2.10.5
repoze.lru==0.6
repoze.sendmail==4.4.1
requests-oauthlib==0.8.0
requests==2.18.1
rsa==3.4.2
scikit-learn==0.18.2
scipy==0.19.1
selenium==3.4.3
simplejson==3.11.1
six==1.10.0
smart-open==1.5.3
snowballstemmer==1.2.1
-e git+https://github.com/assembl/sockjs-tornado.git@conflicting_localhost#egg=sockjs_tornado
SPARQLWrapper==1.8.0
sphinxcontrib-httpdomain==1.5.0
sphinxcontrib-zopeext==0.2.1
Sphinx==1.6.2
# we want the latest
#splinter==0.7.5
-e git+https://github.com/cobrateam/splinter.git#egg=splinter
SQLAlchemy==1.1.11
-e git+https://github.com/conversence/sqlalchemy-rdfbridge.git#egg=sqlalchemy-rdfbridge
sqlalchemy-schemadisplay==1.3
sqlparse==0.2.3
-e git+https://github.com/assembl/sqltap.git@assembl#egg=sqltap
supervisor==3.3.2
tornado==4.5.1
traitlets==4.3.2
transaction==2.1.2
translationstring==1.3
Unidecode==0.04.21
uritemplate==3.0.0
uwsgitop==0.10
venusian==1.1.0
waitress==1.0.2
watchdog==0.8.3
webencodings==0.5.1
WebOb==1.7.2
WebTest==2.0.27
Werkzeug==0.12.2
zope.component==4.3.0
zope.deprecation==4.2.0
zope.event==4.2.0
zope.interface==4.4.2
zope.sqlalchemy==0.7.7
# To be watched:
# https://github.com/elliotpeele/pyramid_oauth2_provider
